% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{addFlatCoordinatesSubcortical}
\alias{addFlatCoordinatesSubcortical}
\title{Add new coordinates to subcortical annotated maps - which allow for flat plotting}
\usage{
addFlatCoordinatesSubcortical(annotBwas, hemi)
}
\arguments{
\item{annotBwas}{Annotated subcortical brain map}

\item{hemi}{Hemisphere ("lh" or "rh")}
}
\value{
Annotated subcortical brain map with additional coordiantes
}
\description{
This function reads in an annotated subcortical brain association map
It add additional sets of coordinates for flat plotting
}
