% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{annot3D}
\alias{annot3D}
\title{Pixel level 3D Structure annotations of the embryonic 11.5 day old mouse brain}
\format{3D array}
\usage{
annot3D
}
\description{
Pixel level 3D Structure annotations of the embryonic 11.5 day old mouse brain
}
\references{
\url{http://developingmouse.brain-map.org/}
}
\keyword{datasets}

