% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{BrainMapAnnotAndManhattanPlot}
\alias{BrainMapAnnotAndManhattanPlot}
\title{Manhattan plot of a brain association map}
\usage{
BrainMapAnnotAndManhattanPlot(
  inputPath,
  bwasFile,
  yMax,
  phenotypeLabel,
  signifThreshold,
  outputPath,
  qqPlot = T
)
}
\arguments{
\item{inputPath}{path (folder) to the raw brain association maps (outputs of OSCA)}

\item{bwasFile}{name of the brain association map}

\item{yMax}{maximum of the y-axis in the Manhattan plot}

\item{phenotypeLabel}{Label of the phenotype - used for plotting}

\item{signifThreshold}{pvalue significance threshold used to account for multiple testing}

\item{outputPath}{path where the outputs will be written}

\item{qqPlot}{True/False: should a qqplot also be created?}
}
\value{
Manhattan plot as well as annotated brain association maps (full and subset of significant vertices)
}
\description{
This function takes in raw brain association maps, formats it and writes plot
}
