% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{BrainMapManhattanPlot}
\alias{BrainMapManhattanPlot}
\title{Manhattan plot of a brain association map}
\usage{
BrainMapManhattanPlot(
  inputPath,
  bwasFile,
  yMax,
  phenotypeLabel,
  signifThreshold,
  outputPath,
  qqPlot
)
}
\arguments{
\item{inputPath}{path (folder) to the raw brain association maps (outputs of OSCA)}

\item{bwasFile}{name of the brain association map}

\item{yMax}{maximum of the y-axis}

\item{phenotypeLabel}{Label of the phenotype - for plotting}

\item{signifThreshold}{significance threshold}

\item{outputPath}{path where the outputs will be written}

\item{qqplot}{True/False: should a qqplot also be created?}
}
\value{
Manhattan plot as well as formatted brain association map (full and subsetted for significant vertices)
}
\description{
This function takes in raw brain association maps, formats it and writes plot
}
