% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{brainmapr}
\alias{brainmapr}
\alias{brainmapr-package}
\title{brainmapr}
\description{
R package to infer spatial location of neuronal subpopulations within the developing mouse brain by integrating single-cell RNA-seq data with in situ RNA patterns from the Allen Developing Mouse Brain Atlas
More extensive tutorials are available at \url{https://github.com/JEFworks/brainmapr}.
}
\author{
Jean Fan \email{jeanfan@fas.harvard.edu}
}

