% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{bwasAddCoordinatesSubcortical}
\alias{bwasAddCoordinatesSubcortical}
\title{Add coordinates to formatted association brain map (Subortical)}
\usage{
bwasAddCoordinatesSubcortical(bwasFormatted, hemi, moda)
}
\arguments{
\item{bwasFormatted}{Formatted brain association map (created using formatBWAScortical)}

\item{hemi}{Hemisphere ("lh" or "rh")}

\item{moda}{modality ("LogJacs" or "thick")}
}
\value{
An enriched brain association maps which contains spatial coordinates
}
\description{
This function opens and format brain association maps to include coordinates (useful for plotting)
The brain association map requires a ProbeID column, which contains the name of vertices/voxels.
It opens the relevant atlas file to merge with the association map.
}
