% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterIdentification}
\alias{clusterIdentification}
\title{Identify and describe all significant clusters from a brain association maps}
\usage{
clusterIdentification(
  outputPath,
  bwas,
  moda,
  signifThreshold,
  KNearestNeighbours,
  phenotypeLabel
)
}
\arguments{
\item{outputPath}{path to the output/writing directory}

\item{bwas}{Formatted brain association map}

\item{moda}{Measurement type ("area" or "thickness" for cortical measurements; "LogJacs" or "thick" for subcortical)}

\item{signifThreshold}{Significance threshold}

\item{KNearestNeighbours}{Number of nearest neighbors to add to the cluster at each iterative step}

\item{phenotypeLabel}{Label of the phenotype for enhanced plotting}
}
\value{
Summary file of significant clusters and their properties
}
\description{
This function reads a formatted brain association map (with X Y Z coordinates).
}
