% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{combineCorticalSubcorticalPlots}
\alias{combineCorticalSubcorticalPlots}
\title{Combine cortical and subcortical figures to create a multi-panel plot}
\usage{
combineCorticalSubcorticalPlots(path, bwasFile, pathToLegendBar)
}
\arguments{
\item{path}{path to working directory (input and output folder)}

\item{bwasFile}{Variable name (used to name the clustersAndCoordinates file)}

\item{pathToLegendBar}{Path to the manually created legend bar}
}
\value{
Outside and Inside snapshots of the surfaces.
}
\description{
This function uses the cortical ans subcortical figures created using the specific functions
}
