% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{combineCorticalSubcorticalPlots_ROI}
\alias{combineCorticalSubcorticalPlots_ROI}
\title{Combine ROI figures to create a multi-panel plot of ROI associations with legend}
\usage{
combineCorticalSubcorticalPlots_ROI(
  inputPath,
  bwasFile,
  pathToLegendBar,
  outputPath,
  style = "smoothwm"
)
}
\arguments{
\item{inputPath}{path (folder) to the cortical and subcortical plots}

\item{bwasFile}{name of bwas file (used in naming of cortical and subcortical plots)}

\item{pathToLegendBar}{Path to the legend bar, created with createLegendBar()}

\item{outputPath}{folder where the outputs will be written}

\item{style}{style for plotting, Possible options: "orig", "pial", "sphere", "inflated",  "inflated_pre",  "pial_semi_inflated", "white", "smoothwm" (default)}
}
\value{
A combined plot with cortical and subcortical surface plots and legendbar
}
\description{
This function uses the cortical and subcortical figures created using the specific functions
}
