% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{combineSubcorticalPlots}
\alias{combineSubcorticalPlots}
\title{Combine subcortical figures to create a multi-panel plot}
\usage{
combineSubcorticalPlots(inputPath, bwasFile, pathToLegendBar, outputPath)
}
\arguments{
\item{inputPath}{path (folder) to the cortical and subcortical plots}

\item{bwasFile}{name of bwas file (used in naming of cortical and subcortical plots)}

\item{pathToLegendBar}{Path to the legend bar, created with createLegendBar()}

\item{outputPath}{folder where the outputs will be written}
}
\value{
A combined plot with cortical and subcortical surface plots and legendbar
}
\description{
This function uses the subcortical figures created using the specific functions
}
