% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{example_gannot3D}
\alias{example_gannot3D}
\title{Voxel level 3D Structure annotations of the embryonic 11.5 day old mouse brain}
\format{3D array}
\usage{
example_gannot3D
}
\description{
Voxel level 3D Structure annotations of the embryonic 11.5 day old mouse brain
}
\references{
\url{http://developingmouse.brain-map.org/}
}
\keyword{datasets}
