% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{example_structureID}
\alias{example_structureID}
\title{Parsed StructureGraph for the developing mouse brain. Contains structure IDs and names in a hierarchical format.}
\format{3D array}
\usage{
example_structureID
}
\description{
Parsed StructureGraph for the developing mouse brain. Contains structure IDs and names in a hierarchical format.
}
\references{
\url{http://developingmouse.brain-map.org/}
}
\keyword{datasets}
