% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{formatBWASsubcortical}
\alias{formatBWASsubcortical}
\title{Open and format association brain map (Subcortical)}
\usage{
formatBWASsubcortical(BWASsumstat)
}
\arguments{
\item{BWASsumstat}{Brain association map (OSCA format)}
}
\value{
An enriched brain association maps which contains more vertex information
}
\description{
This function opens and format brain association maps to include coordinates and ROI (ENIGMA-shape atlas).
The brain association map requires a ProbeID column, which contains the name of vertices/voxels.
It opens the relevant atlas file to merge with the association map.
}
