% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{genePlot}
\alias{genePlot}
\title{Plot total expression of a set of genes}
\usage{
genePlot(gl, expmat, gannot, t = 0, weights = rep(1, length(gl)),
  plot = F)
}
\arguments{
\item{gl}{Gene list}

\item{expmat}{3D volume of expression}

\item{gannot}{3D annotation of volume with structure IDs}

\item{weights}{List of relative weights corresponding to gene list; default: weighted equally}

\item{plot}{Boolean of whether to make 3D plot}
}
\value{
exp3D 3D volume restricted to weighted gene expression set of interest
}
\description{
Plot total expression of a set of genes
}

