% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{genePlotWeightedComp2}
\alias{genePlotWeightedComp2}
\title{Relative expression of two groups using ratios}
\usage{
genePlotWeightedComp2(gl1, weights1, gl2, weights2, expmat, gannot, plot = F)
}
\arguments{
\item{gl1}{List of differentially expressed genes upregulated in cell group 1}

\item{weights1}{List of weights corresponding to genes in \code{gl1}}

\item{gl2}{List of differentially expressed genes upregulated in cell group 2}

\item{weights2}{List of weights corresponding to genes in \code{gl2}}

\item{expmat}{3D volume of expression}

\item{gannot}{3D annotation of volume with structure IDs}

\item{plot}{Boolean of whether to make 3D plot}
}
\value{
exp3D 3D volume of weighted expression for gene lists of interest
}
\description{
Relative expression of two groups using ratios
}
\keyword{comparison,}
\keyword{placement}

