% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{getChildren}
\alias{getChildren}
\title{A recursive way to get all the children of a Structure given its ID. Used by \code{\link{getStructureIds}}}
\usage{
getChildren(nestedList, sid)
}
\arguments{
\item{nestedList}{A JSON derived nested list of structure relationships from Allen Brain Atlas's structure graph}

\item{sid}{The Structure's ID}
}
\value{
cids The Structure children's IDs
}
\description{
A recursive way to get all the children of a Structure given its ID. Used by \code{\link{getStructureIds}}
}
\examples{
data(structureID)
getChildren(structureID, 15903)
}
\keyword{browse,}
\keyword{helper}
\keyword{search,}

