% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{getCorrelationRange}
\alias{getCorrelationRange}
\title{Scan association file to return the range of the association statistics (for the significant vertices)}
\usage{
getCorrelationRange(inputPath, bwasFile, variancePheno, signifThreshold)
}
\arguments{
\item{inputPath}{path (folder) to the raw brain association maps (outputs of OSCA)}

\item{bwasFile}{name of the brain association map}

\item{variancePheno}{variance of the phenotypes (to estimate correlation)}

\item{signifThreshold}{pvalue significance threshold used to account for multiple testing}
}
\value{
Range of correlation coefficients, useful to set appropriate legend bar and cor range in plotting function
}
\description{
This function takes in raw brain association maps, and returns the range of association, which can be used to inform plotting
}
