% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{getId}
\alias{getId}
\title{A recursive way to get structure IDs and names of structure given an exact name string. Used by \code{\link{getStructureIds}}}
\usage{
getId(nestedList, name)
}
\arguments{
\item{nestedList}{A JSON derived nested list of Structure relationships from Allen Brain Atlas's structure graph}

\item{name}{Structure's name}
}
\value{
cids Structure ID
}
\description{
A recursive way to get structure IDs and names of structure given an exact name string. Used by \code{\link{getStructureIds}}
}
\examples{
data(structureID)
getId(structureID, 'pallium')
}
\keyword{browse,}
\keyword{helper}
\keyword{search,}

