% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{getIds}
\alias{getIds}
\title{A recursive way to get IDs and names of structure given a name substring}
\usage{
getIds(nestedList, name)
}
\arguments{
\item{nestedList}{A JSON derived nested list of Structure relationships from Allen Brain Atlas's StructureGraph}

\item{name}{Brain structure name substring}
}
\value{
ids A dictionary of Structures that contain the name with their IDs
}
\description{
In the Allen Brain Atlas, a Structure represents a neuroanatomical region of interest. Structures are grouped into Ontologies and organized in a hierarchy or StructureGraph. The StructureGraph for the Developing Mouse Brain Atlas is provided here under data/structureIDRda. This function allows you to browse Structures by name and ID given a name substring.
}
\examples{
data(structureID)
getIds(structureID, 'pallium')
}
\keyword{browse,}
\keyword{search}

