% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{getStructureIds}
\alias{getStructureIds}
\title{Get IDs associated with a Structure given the exact Structure name}
\usage{
getStructureIds(nestedList, name)
}
\arguments{
\item{nestedList}{A JSON derived nested list of structure relationships from Allen Brain Atlas's StructureGraph}

\item{name}{Structure's name}
}
\value{
cids The Structure's ID and all its children's IDs
}
\description{
In the Allen Brain Atlas, a Structure represents a neuroanatomical region of interest. Structures are grouped into Ontologies and organized in a hierarchy or StructureGraph. With the exception of the "root" structure, each Structure has one parent and denotes a "part-of" relationship. Due to this relationship, a Structure's total volume is denoted by both its own IDs and also that of its children's IDs. The StructureGraph for the Developing Mouse Brain Atlas is provided here under data/structureID.rda. This function allows you to obtain all IDs associated with a Structure given its exact name.
}
\examples{
data(structureID)
getStructureIds(structureID, 'pallium')
}
\keyword{browse,}
\keyword{search}

