% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumr2.R
\name{getWeights}
\alias{getWeights}
\title{Extract weights from weighted least square during sumR2 regression}
\usage{
getWeights(
  ld_score,
  ld_size,
  chi2,
  sample_size,
  blocks = 200,
  intercept = NULL,
  chi2_thr1 = 30,
  chi2_thr2 = Inf
)
}
\arguments{
\item{ld_score}{ld_score}

\item{ld_size}{ld_size}

\item{chi2}{chi2}

\item{sample_size}{sample_size}

\item{blocks}{blocks}

\item{intercept}{intercept}

\item{chi2_thr1}{chi2_thr1}

\item{chi2_thr2}{chi2_thr2}
}
\value{
weights from SumR2 regression
}
\description{
Extract weights from weighted least square during sumR2 regression
}
