% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{identifyClustersBWAS}
\alias{identifyClustersBWAS}
\title{Run all the cluster identification from a raw brain map}
\usage{
identifyClustersBWAS(inputPath, bwasFile, outputPath, signifThreshold)
}
\arguments{
\item{inputPath}{path (folder) to the raw brain association maps (outputs of OSCA)}

\item{bwasFile}{name of the brain association map (without extension)}

\item{outputPath}{path to output folder}

\item{signifThreshold}{significance threshold}
}
\value{
Summary file of significant clusters and their properties
}
\description{
This function combines, reading and formatting fuctions as well as the clustering ones to perform all the analyses for you.
}
