% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{mat}
\alias{mat}
\title{Matrix of gene expression where each row corresponds to a gene and columns correspond to 3D locations in space}
\format{2D array (matrix)}
\usage{
mat
}
\description{
Matrix of gene expression where each row corresponds to a gene and columns correspond to 3D locations in space
}
\references{
\url{http://developingmouse.brain-map.org/}
}
\keyword{datasets}

