% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCortical}
\alias{plotCortical}
\title{Cortical brain plot}
\usage{
plotCortical(path, bwasFile, pathPhenotypeFile)
}
\arguments{
\item{path}{path to working directory (input and output folder)}

\item{bwasFile}{name of the brain association map (without extension)}

\item{pathPhenotypeFile}{Path to the variable files (e.g. variable.phen) used get variance of the phenotype}
}
\value{
Outside and Inside snapshots of the surfaces.
}
\description{
This function reads in an annotated brain association map (cluster and coordinates).
It produces snapshots of the brain surfaces.
The function opens the phenotype file in order to calculate the variance of the phenotype, used to convert association betas into correlation coefficients.
}
