% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotCorticalAtlasToAtlasGif}
\alias{plotCorticalAtlasToAtlasGif}
\title{Cortical brain plots for GIF - across styles of plotting}
\usage{
plotCorticalAtlasToAtlasGif(
  inputPath,
  bwasFile,
  variancePheno,
  signifThreshold,
  moda,
  hemi,
  outputPath,
  nbImagesForGif,
  correlationRange = c(-1, 1),
  faster,
  leftOrRightView,
  styleStart,
  styleEnd
)
}
\arguments{
\item{inputPath}{path (folder) to the raw brain association maps}

\item{bwasFile}{name of the brain association map}

\item{variancePheno}{Variance of the phenotype (used to standardise the effect sizes into correlations)}

\item{signifThreshold}{pvalue significance threshold}

\item{moda}{modality to plot ("thickness" or "area")}

\item{hemi}{hemisphere to plot ("lh" or "rh")}

\item{outputPath}{path where the outputs will be written. Absolute path may be required to produce GIF.}

\item{nbImagesForGif}{Number of png images to generate for the gif. The larger the smoother the gif, but the longer it takes to generate. Please use a multiple of 6 if you want the GIF to be created automatically.}

\item{correlationRange}{range of the correlation coefficients (for improved colors) - default is (-1; 1)}

\item{faster}{downsample mesh for faster plotting (e.g. testing)}

\item{leftOrRightView}{direction of the view (e.g. for left hemisphere left view corresponds to outside view)}

\item{styleStart}{style for plotting, which will be the starting style of the GIF. Possible options: "orig", "pial", "sphere", "inflated",  "inflated_pre",  "pial_semi_inflated", "white"}

\item{styleEnd}{style for plotting, which will be the ending style of the GIF. Possible options: "orig", "pial", "sphere", "inflated",  "inflated_pre",  "pial_semi_inflated", "white"}
}
\value{
Several snapshots of the rotating cortex and a GIF
}
\description{
This function reads in a brain association map.
It produces snapshots of the brain surfaces that progressively go from one surface style to another (e.g. from pial to inflated).
The function needs the variance of the phenotype, in order to convert association betas into correlation coefficients.
}
