% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotEnergy}
\alias{plotEnergy}
\title{Helper function for 3D plotting. Used by structurePlot() and genePlot()}
\usage{
plotEnergy(exp, x, y, z)
}
\arguments{
\item{exp}{vector representation of volume}

\item{x}{x-dimension of volume}

\item{y}{y-dimension of volume}

\item{z}{z-dimension of volume}
}
\description{
Helper function for 3D plotting. Used by structurePlot() and genePlot()
}

