% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotProjectionXray}
\alias{plotProjectionXray}
\title{\code{\link{plotProjection}} with colors set to look more like an x-ray}
\usage{
plotProjectionXray(mat3D, t = 0, add = F)
}
\arguments{
\item{mat3D}{3D volume}

\item{t}{Threshold used to remove noise}

\item{add}{Boolean whether to overlay onto existing plot}
}
\description{
\code{\link{plotProjection}} with colors set to look more like an x-ray
}
\keyword{modification}
\keyword{plot,}

