% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotROIs_corticalAndSubcortical}
\alias{plotROIs_corticalAndSubcortical}
\title{ROI (Region of Interest) plot}
\usage{
plotROIs_corticalAndSubcortical(
  inputPath,
  bwasFile,
  variancePheno,
  outputPath,
  signifThreshold,
  correlationRange = NULL,
  betaCol,
  seCol,
  pvalCol,
  style = "smoothwm"
)
}
\arguments{
\item{inputPath}{path (folder) to the ROI level brain association maps}

\item{bwasFile}{name of the brain association map}

\item{variancePheno}{Variance of the phenotype (used to standardise the effect sizes into correlations)}

\item{outputPath}{path where the outputs will be written}

\item{signifThreshold}{pvalue significance threshold}

\item{correlationRange}{range of the correlation coefficients (for improved colors) - if not specified, it is inferred from the data}

\item{betaCol}{column name corresponding to beta or effect size to plot}

\item{seCol}{column name corresponding to the SE of effect size}

\item{pvalCol}{column name corresponding to the pvalue associated with effect size}

\item{style}{style for plotting, Possible options: "orig", "pial", "sphere", "inflated",  "inflated_pre",  "pial_semi_inflated", "white", "smoothwm" (default)}
}
\value{
Several snapshots of the cortical surfaces and subcortical stuctures and intermediate text files used for plotting. Last file created is the full figure.
}
\description{
This function reads in a brain association map at the ROI level
It produces snapshots of the associations with ROI-level cortical thickness, cortical surface and subcortical volumes
Then it combines all plots in a summary plot - and adds the legend
Expected ROIs are the FreeSurfer derived measurements, that use the standard naming used in FreeSurfer and ENIGMA
The function needs the variance of the phenotype, in order to convert association betas into correlation coefficients.
}
