% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotSlice}
\alias{plotSlice}
\title{Plot slice of 3D volume}
\usage{
plotSlice(mat3D, slice, col = colorRampPalette(c("white", "black", "red",
  "yellow"), space = "Lab")(100), t = 0, add = F)
}
\arguments{
\item{mat3D}{3D volume}

\item{slice}{Index of slice; limited to dimensions of mat3D}

\item{col}{Color}

\item{t}{Threshold used to remove noise}

\item{add}{Boolean whether to overlay onto existing plot}
}
\description{
Plot slice of 3D volume
}
\examples{
data(vol3D)
plotSlice(vol3D, 75, t=8)
}
\keyword{plot}

