% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotSliceComp}
\alias{plotSliceComp}
\title{Compare 2 regions}
\usage{
plotSliceComp(mat3D, slice, col = colorRampPalette(c("green", "blue", "black",
  "red", "yellow"), space = "Lab")(100), add = F)
}
\description{
Compare 2 regions
}

