% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSubcortical}
\alias{plotSubcortical}
\title{Subcortical brain plot}
\usage{
plotSubcortical(
  inputPath,
  bwasFile,
  variancePheno,
  outputPath,
  signifThreshold,
  correlationRange = c(-1, 1)
)
}
\arguments{
\item{inputPath}{path (folder) to the raw brain association maps (outputs of identifyClustersBWAS())}

\item{bwasFile}{name of the brain association map}

\item{variancePheno}{Variance of the phenotype (used to standardise the effect sizes into correlations)}

\item{outputPath}{path where the outputs will be written}

\item{signifThreshold}{pvalue significance threshold used to account for multiple testing}

\item{correlationRange}{range of the correlation coefficients (for improved colors) - default is (-1; 1)}
}
\value{
Outside and Inside snapshots of the surfaces.
}
\description{
This function reads in an annotated brain association map (cluster and coordinates).
It produces snapshots of the brain surfaces.
The function opens the phenotype file in order to calculate the variance of the phenotype, used to convert association betas into correlation coefficients.
}
