% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSubcortical_flat}
\alias{plotSubcortical_flat}
\title{Better rendition of the subcortical brain plots}
\usage{
plotSubcortical_flat(
  inputPath,
  bwasFile,
  variancePheno,
  signifThreshold,
  outputPath,
  correlationRange = c(-1, 1)
)
}
\arguments{
\item{inputPath}{path (folder) to the raw brain association maps (outputs of identifyClustersBWAS())}

\item{bwasFile}{name of the brain association map}

\item{variancePheno}{Variance of the phenotype (used to standardise the effect sizes into correlations)}

\item{signifThreshold}{pvalue significance threshold used to account for multiple testing}

\item{outputPath}{path where the outputs will be written}

\item{correlationRange}{range of the correlation coefficients (for improved colors) - default is (-1; 1)}
}
\value{
Outside and Inside snapshots of the surfaces.
}
\description{
This function reads in an annotated brain association map (cluster and coordinates).
It produces snapshots of the brain surfaces.
The function opens the phenotype file in order to calculate the variance of the phenotype, used to convert association betas into correlation coefficients.
}
