% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSubcortical_FlatGIF}
\alias{plotSubcortical_FlatGIF}
\title{Subcortical GIF flat - rotation of each subcortical structure from the flat format}
\usage{
plotSubcortical_FlatGIF(
  inputPath,
  bwasFile,
  variancePheno,
  outputPath,
  hemi,
  moda,
  signifThreshold,
  nbImagesForGif,
  leftOrRightView,
  correlationRange = c(-1, 1)
)
}
\arguments{
\item{inputPath}{path (folder) to the raw brain association maps}

\item{bwasFile}{name of the brain association map}

\item{variancePheno}{Variance of the phenotype (used to standardise the effect sizes into correlations)}

\item{outputPath}{path where the outputs will be written}

\item{hemi}{hemisphere to plot ("lh" or "rh")}

\item{moda}{modality to plot ("thick" or "LogJacs")}

\item{signifThreshold}{pvalue significance threshold}

\item{nbImagesForGif}{Number of png images to generate for the gif. The larger the smoother the gif, but the longer it takes to generate.}

\item{leftOrRightView}{The side from which the snapshots will be taken}

\item{correlationRange}{range of the correlation coefficients (for improved colors) - default is (-1; 1)}
}
\value{
Several snapshots of the surfaces.
}
\description{
This function reads in a brain association map
It produces snapshots of the subcortical brain surfaces, with a slight rotation angle, which can be used to make a GIF
The function needs the variance of the phenotype, in order to convert association betas into correlation coefficients.
}
