% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{significantClusterAroundVertex}
\alias{significantClusterAroundVertex}
\title{Identify significant cluster around a (significant) vertex of interest}
\usage{
significantClusterAroundVertex(
  bwas,
  signifThreshold,
  targetVertex,
  nbNearestNeighbours
)
}
\arguments{
\item{bwas}{Formatted brain association map}

\item{signifThreshold}{Significance threshold}

\item{targetVertex}{Name of the vertex of interest}

\item{nbNearestNeighbours}{Number of nearest neighbors to add to the cluster at each iterative step}
}
\value{
An updated brain association map with cluster membership
}
\description{
This function reads a formatted brain association map (with X Y Z coordinates).
From the target vertex, the algorithm identifies neighboring vertices that are also significant.
}
