% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{structureID}
\alias{structureID}
\title{Parsed StructureGraph for the developing mouse brain. Contains structure IDs and names in a hierarchical format.}
\format{3D array}
\usage{
structureID
}
\description{
Parsed StructureGraph for the developing mouse brain. Contains structure IDs and names in a hierarchical format.
}
\references{
\url{http://developingmouse.brain-map.org/}
}
\keyword{datasets}

