% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{structurePlot}
\alias{structurePlot}
\title{Plot the volume or expression within a Structure}
\usage{
structurePlot(cids, vol, annot, plot = F)
}
\arguments{
\item{cids}{Structure ids}

\item{vol}{3D volume of expression values}

\item{annot}{3D annotation of volume with structure IDs}

\item{plot}{Boolean of whether to make 3D plot}
}
\value{
tvol 3D volume restricted to just structure of interest
}
\description{
Plot the volume or expression within a Structure
}
\examples{
data(structureID)
data(vol3D)
data(annot3D)
cids <- getStructureIds(structureID, 'pallium')
sect3D <- structurePlot(cids, vol3D, annot3D)  # For whole structure
gpsect3D <- structurePlot(cids, array(mat['Dcx',], dim=dim(gannot3D)), gannot3D)  # For a particular gene
}
\keyword{plot}

