% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumr2.R
\name{sumR2_regression_bivariate}
\alias{sumR2_regression_bivariate}
\title{SumR2 regression - estimate grey-matter correlation (rGM)}
\usage{
sumR2_regression_bivariate(
  inputPath = inputPath,
  bwasFile = bwasFile,
  refPanel = c("ADNI1", "ADNI2GO3", "AIBL"),
  nblock = 200,
  chi2Threshold = 80,
  bwasSampleSize = c("NMISS", "NMISS"),
  varConstrained = TRUE
)
}
\arguments{
\item{inputPath}{Paths (folders) to the raw brain association maps (outputs of OSCA)}

\item{bwasFile}{Brain association maps in (OSCA format)}

\item{refPanel}{Reference panel for the sumR2 -  (default UK Biobank, options include ADNI1, ADNI2GO3, AIBL). If several reference panels are listed, results using each sumR2 is returned.}

\item{nblock}{Number of blocks used in jacknife to estimate SE  - (default 200)}

\item{chi2Threshold}{Chi2 threshold to exclude vertices with outlying large association - (default 80)}

\item{bwasSampleSize}{Optional : If character, the name of the columns in bwas file that contains the sample size (default "NMISS"). If numeric, the sample sizes.}

\item{varConstrained}{(TRUE/FALSE, default = T) whether variance components estimates should be constrained to be between 0 and 1}
}
\value{
Morphometricity, SE, confidence intervals and pvalue, SumR2 intercept and SE.
}
\description{
This function reads two brain association map
It annotates the association maps to add SumR2 pre-calculated on reference samples
Then performs sumR2 regression to estimate rGM
}
