% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumr2.R
\name{WEIGHTS}
\alias{WEIGHTS}
\title{Heteroscedasticity and overcounting weights}
\usage{
WEIGHTS(pred, w_ld)
}
\arguments{
\item{pred}{pred}

\item{w_ld}{w_ld}
}
\description{
Heteroscedasticity and overcounting weights
}
