% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumr2.R
\name{wlm}
\alias{wlm}
\title{Equivalent to stats::lm.wfit(cbind(1, x), y, w)}
\usage{
wlm(x, y, w)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{w}{w}
}
\description{
Equivalent to stats::lm.wfit(cbind(1, x), y, w)
}
