% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumr2.R
\name{wlm_no_int}
\alias{wlm_no_int}
\title{Equivalent to stats::lm.wfit(as.matrix(x), y, w)}
\usage{
wlm_no_int(x, y, w)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{w}{w}
}
\description{
Equivalent to stats::lm.wfit(as.matrix(x), y, w)
}
