% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{BrainMapManhattanPlot}
\alias{BrainMapManhattanPlot}
\title{Manhattan plot of a brain association map}
\usage{
BrainMapManhattanPlot(path, bwasFile, yMax, phenotypeLabel, signifThreshold)
}
\arguments{
\item{path}{path (folder) to the raw brain association maps (outputs of OSCA)}

\item{bwasFile}{name of the brain association map}

\item{yMax}{maximum of the y-axis}

\item{phenotypeLabel}{Label of the phenotype - for plotting}

\item{signifThreshold}{significance threshold}
}
\value{
Manhattan plot as well as formatted brain association map (full and subsetted for significant vertices)
}
\description{
This function takes in raw brain association maps, formats it and writes plot
}
