% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{combineCorticalSubcorticalPlots}
\alias{combineCorticalSubcorticalPlots}
\title{Combine cortical and subcortical figures to create a multi-panel plot}
\usage{
combineCorticalSubcorticalPlots(
  inputPath,
  bwasFile,
  pathToLegendBar,
  outputPath
)
}
\arguments{
\item{inputPath}{path (folder) to the cortical and subcortical plots}

\item{bwasFile}{Variable name (used to name the clustersAndCoordinates file)}

\item{pathToLegendBar}{Path to the manually created legend bar}

\item{outputPath}{path where the outputs will be written}
}
\value{
Outside and Inside snapshots of the surfaces.
}
\description{
This function uses the cortical ans subcortical figures created using the specific functions
}
