% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{formatBWAScortical}
\alias{formatBWAScortical}
\title{Open and format association brain map (Cortical)}
\usage{
formatBWAScortical(BWASsumstat, hemi, mod)
}
\arguments{
\item{BWASsumstat}{Brain association map (OSCA format)}

\item{hemi}{Hemisphere ("lh" or "rh")}

\item{mod}{modality ("area" or "thickness")}
}
\value{
An enriched brain association maps which contains more vertex information
}
\description{
This function opens and format brain association maps to include coordinates and ROI (Desikan atlas).
The brain association map requires a ProbeID column, which contains the name of vertices/voxels.
It opens the relevant atlas file to merge with the association map.
}
