% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{formatBWASsubcortical}
\alias{formatBWASsubcortical}
\title{Open and format association brain map (Subcortical)}
\usage{
formatBWASsubcortical(BWASsumstat, hemi, mod)
}
\arguments{
\item{BWASsumstat}{Brain association map (OSCA format)}

\item{hemi}{Hemisphere ("lh" or "rh")}

\item{mod}{modality ("LogJacs" or "thick")}
}
\value{
An enriched brain association maps which contains more vertex information
}
\description{
This function opens and format brain association maps to include coordinates and ROI (from ENIGMA-shape atlas). All this information is required to produce Manhattan plots.
The brain association map requires a ProbeID column, which contains the name of vertices/voxels.
It opens the relevant atlas file (incl. in the R package) to annotate the association map.
}
