% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{genePlotWeightedComp}
\alias{genePlotWeightedComp}
\title{Relative expression of two groups using absolute differences}
\usage{
genePlotWeightedComp(gl1, weights1, gl2, weights2, expmat, gannot, plot = F)
}
\description{
Relative expression of two groups using absolute differences
}

