% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{genePlotWeightedComp2}
\alias{genePlotWeightedComp2}
\title{Relative expression of two groups using ratios}
\usage{
genePlotWeightedComp2(gl1, weights1, gl2, weights2, expmat, gannot, plot = F)
}
\description{
Relative expression of two groups using ratios
}

