% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumr2.R
\name{getWeights}
\alias{getWeights}
\title{Extract weights from weighted least square during sumR2 regression}
\usage{
getWeights(
  ld_score,
  ld_size,
  chi2,
  sample_size,
  blocks = 200,
  intercept = NULL,
  chi2_thr1 = 30,
  chi2_thr2 = Inf
)
}
\description{
Extract weights from weighted least square during sumR2 regression
}
