% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsSimpler.R
\name{plotCortical}
\alias{plotCortical}
\title{Cortical brain plot}
\usage{
plotCortical(inputPath, bwasFile, pathPhenotypeFile, outputPath)
}
\arguments{
\item{inputPath}{path (folder) to the brain association maps (outputs of identifyClustersBWAS())}

\item{bwasFile}{name of the brain association map (without extension)}

\item{pathPhenotypeFile}{Path to the variable files (e.g. variable.phen) used get variance of the phenotype}

\item{outputPath}{path where the outputs will be written}
}
\value{
Outside and Inside snapshots of the surfaces.
}
\description{
This function reads in an annotated brain association map (cluster and coordinates).
It produces snapshots of the brain surfaces.
The function opens the phenotype file in order to calculate the variance of the phenotype, used to convert association betas into correlation coefficients.
}
