% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotProjectionXray}
\alias{plotProjectionXray}
\title{Set the colors to look more like an x-ray}
\usage{
plotProjectionXray(mat3D, col = colorRampPalette(c("white", "black"), space =
  "Lab")(100), t = 0, add = F)
}
\description{
Set the colors to look more like an x-ray
}

