% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotSliceComp}
\alias{plotSliceComp}
\title{Plot slice of 3D volume without thresholding and allowing for negative expression values. Used for comparing spatial expression of upregulated vs. downregulated gene sets.}
\usage{
plotSliceComp(mat3D, slice, col = colorRampPalette(c("green", "blue", "black",
  "red", "yellow"), space = "Lab")(100), add = F)
}
\arguments{
\item{mat3D}{3D volume}

\item{slice}{Index of slice; limited to dimensions of mat3D}

\item{col}{Color}

\item{add}{Boolean whether to overlay onto existing plot}
}
\description{
Plot slice of 3D volume without thresholding and allowing for negative expression values. Used for comparing spatial expression of upregulated vs. downregulated gene sets.
}
\examples{
data(vol3D)
plotSliceComp(vol3D, 75)
}
\keyword{plot}

