% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/functions.R
\name{plotSliceXray}
\alias{plotSliceXray}
\title{\code{\link{plotSlice}} with colors set to look more like an x-ray}
\usage{
plotSliceXray(mat3D, slice, t = 0, add = F)
}
\arguments{
\item{mat3D}{3D volume}

\item{slice}{Index of slice; limited to dimensions of mat3D}

\item{t}{Threshold used to remove noise}

\item{add}{Boolean whether to overlay onto existing plot}
}
\description{
\code{\link{plotSlice}} with colors set to look more like an x-ray
}
\examples{
data(vol3D)
plotSliceXray(vol3D, 75, t=8)
}
\keyword{modification}
\keyword{plot,}

