% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{qqplot}
\alias{qqplot}
\title{QQplot of an association map (or any vector of pvalues)}
\usage{
qqplot(pvector, col = "black", add = F)
}
\arguments{
\item{pvector}{Vector of pvalues}

\item{col}{colour of points}

\item{add}{Add to existing plot (T/F)}
}
\value{
QQplot
}
\description{
This function takes in a vector of pvalues and creates a qqplot (in console)
}
