% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumr2.R
\name{sumR2_regression_univariate}
\alias{sumR2_regression_univariate}
\title{SumR2 regression - estimate morphometricity}
\usage{
sumR2_regression_univariate(
  inputPath,
  bwasFile,
  refPanel,
  nblock = 200,
  chi2Threshold = 80,
  bwasSampleSize = "NMISS",
  outputPath
)
}
\arguments{
\item{inputPath}{path (folder) to the raw brain association maps (outputs of OSCA)}

\item{bwasFile}{Brain association map in (OSCA format)}

\item{refPanel}{Reference panel for the sumR2 -  (default UK Biobank, options include ADNI1, ADNI2GO3, AIBL, AVERAGE). If several reference panels are listed, results using each sumR2 is returned.}

\item{nblock}{Number of blocks used in jacknife to estimate SE  - (default 200)}

\item{chi2Threshold}{Chi2 threshold to exclude vertices with outlying large association - (default 80)}

\item{bwasSampleSize}{Optional : If character, the name of the columns in bwas file that contains the sample size (default "NMISS"). If numeric, the sample size.}

\item{outputPath}{path where the outputs will be written}
}
\value{
Morphometricity, SE, confidence intervals and pvalue, SumR2 intercept and SE.
}
\description{
This function reads a brain association map
It annotates the association map to add SumR2 pre-calculated on reference samples
Then performs sumR2 regression to estimate morphometricity
}
