% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotsSimpler.R
\name{superimposedQQplot}
\alias{superimposedQQplot}
\title{QQplot of several brain association maps}
\usage{
superimposedQQplot(
  inputPaths,
  bwasFiles,
  legendList,
  colourList,
  phenotypeLabel,
  outputPath
)
}
\arguments{
\item{inputPaths}{paths to working directories containing the files BWAS_fullSummary_....csv, created by BrainMapManhattanPlot()}

\item{bwasFiles}{list of brain association maps within each inputPath}

\item{legendList}{List of labels for the different scenarios/folders}

\item{colourList}{List of colours for the different scenarios/folders}

\item{phenotypeLabel}{Label of the phenotype - used for plotting}

\item{outputPath}{path where the output will be written}
}
\value{
QQplot with several distributions
}
\description{
This function reads in a list of brain association maps, which have been formatted and created as part of the Manhattan plot creation (this allows inferring more easily the file names)
}
