% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{superimposedQQplot}
\alias{superimposedQQplot}
\title{QQplot of several brain association maps}
\usage{
superimposedQQplot(
  inputPath,
  scenarioList,
  legendList,
  colourList,
  variableLabel,
  outputPath
)
}
\arguments{
\item{inputPath}{path to working directory containing the outputs of BrainMapManhattanPlot()}

\item{scenarioList}{List of scenarios/folders each containing a brain association map}

\item{legendList}{List of labels for the different scenarios/folders}

\item{colourList}{List of colours for the different scenarios/folders}

\item{variableLabel}{label of the variable}

\item{outputPath}{path where the outputs will be written}
}
\value{
QQplot with several distributions
}
\description{
This function reads in a list of brain association maps, which have been formatted and created as part of the Manhattan plot creation (this allows inferring more easily the file names)
}
